// src/app/models/rifa.ts
export class Rifa {
    rifaId: number;
    statusVisualizacaoId: number;
    statusRifaId: number;
    regulamentoId: number;
    sorteios: any[];
    numeroMin: number;
    numeroMax: number;
    whatsAppContato: string;
    titulo: string;
    preco: number;
    dataFimJogadas: Date;
    descricao: number;
    lancamento: Date;
}